requirements.txt
iniconfig==2.1.0
pygments==2.19.2
soupsieve==2.8
coverage==7.10.7
markupsafe==3.0.3
pluggy==1.6.0
jinja2==3.1.6
packaging==25.0
beautifulsoup4==4.14.2
annotated-types==0.7.0
typing-inspection==0.4.2
typing-extensions==4.15.0
pydantic-core==2.41.4
pydantic==2.12.3
requests==2.32.5
pandas==2.3.3
fastapi==0.118.0
anyio==4.11.0
idna==3.10
sniffio==1.3.1
starlette==0.48.0
pyodbc==5.3.0
types-pytz==2025.2.0.20250809
types-requests==2.32.4.20250913
mypy==1.18.2
pathspec==0.12.1
mypy-extensions==1.1.0
ruff==0.13.2
numpy==2.3.0
filelock==3.19.1
termcolor==3.1.0
pytest==8.2.0
pytest-ruff==0.5
pytest-cov==7.0.0
pytest-pylint==0.21.0
pytest-metadata==3.1.1
pytest-html==4.1.1
pytest-html-merger==0.1.0
pytest-mypy==1.0.1
pytest-sugar==1.1.1
six==1.17.0
tzdata==2025.2
pytz==2025.2
python-dateutil==2.9.0.post0
urllib3==2.5.0
charset-normalizer==3.4.3
certifi==2025.8.3
astroid==4.0.1
tomlkit==0.13.3
isort==6.1.0
dill==0.4.0
platformdirs==4.4.0
mccabe==0.7.0
decorator==5.2.1
scipy==1.16.2
pystoi==0.4.1
mir-eval==0.8.2
pesq==0.0.4
sqlalchemy==2.0.44
einops==0.8.1
cached-property==2.0.1
pb-bss-eval==0.0.2
asteroid==0.7.0
pylint==4.0.2
pylint-plugin-utils==0.9.0
pylint-pytest==1.1.8
pylint-pydantic==0.4.0
pylint-protobuf==0.22.0
taplo==0.9.3
yamlfmt==1.1.1
ruamel.yaml.clib==0.2.14
ruamel.yaml==0.17.40
PyYAML==6.0.3
yamllint==1.37.1
pre-commit==4.3.0
bazel-runfiles==1.6.3
nuitka==2.8.4
ordered-set==4.1.0
zstandard==0.25.0
grpcio==1.76.0
paramiko==4.0.0
psutil==7.1.1
lxml==6.0.2
protobuf==6.33.0
xmltodict==1.0.2
types-protobuf==6.32.1.20250918
